requirements.txt
streamlit
tensorflow>=2.5
pillow
numpy
